plasma config options
