+++ server/app/app_test.go
@@ -1,0 +1,8 @@
+package app
+
+import (
+	"testing"
+)
+
+func TestBadSpendMsg(t *testing.T) {
+}
