--- server/plasmad/app/app_test.go
@@ -1,8 +1,0 @@
-package app
-
-import (
-	"testing"
-)
-
-func TestBadSpendMsg(t *testing.T) {
-}
